+++ androidApp/src/main/java/com/bupware/wedraw/android/logic/retrofit/services/GroupService.kt
@@ -1,7 +1,6 @@
 package com.bupware.wedraw.android.logic.retrofit.services
 
 import com.bupware.wedraw.android.logic.models.Group
-import com.bupware.wedraw.android.logic.models.User
 import retrofit2.Call
 import retrofit2.http.Body
 import retrofit2.http.GET
@@ -24,13 +23,13 @@ interface GroupService {
     fun getGroupByCode(@Path("code") code: String): Call<Group?>
 
     @GET("/weDraw/groups/isFull/{groupID}")
-    fun isGroupFull(@Path("groupID") groupID: Int): Call<Boolean>
+    fun isGroupFull(@Path("groupID") groupID: Long): Call<Boolean>
 
     @POST("/weDraw/groups/create/{name}/{leaderId}")
     fun createGroup(@Path("name") name: String,@Path("leaderId") leaderId: String): Call<String?>
 
     @POST("/weDraw/groups/{userID}/{groupID}")
-    fun insertUsertoUserGroup(@Path("userID") userID: String,@Body groupID: Int): Call<Boolean>
+    fun insertUsertoUserGroup(@Path("userID") userID: String, @Body groupID: Long): Call<Boolean>
 
     @PUT("/weDraw/groups/{id}")
     fun updateGroup(@Path("id") id: Int,@Body group: Group): Call<Boolean>
